refactor: move all database concerns to server app
