+++ apps/game/package.json
@@ -18,7 +18,6 @@
   },
   "dependencies": {
     "@mp/auth": "workspace:*",
-    "@mp/db": "workspace:*",
     "@mp/encoding": "workspace:*",
     "@mp/engine": "workspace:*",
     "@mp/ioc": "workspace:*",
@@ -38,8 +37,7 @@
     "@mp/ws": "workspace:*",
     "pixi.js": "catalog:",
     "solid-js": "catalog:",
-    "uniqolor": "1.1.1",
-    "unique-names-generator": "4.7.1"
+    "uniqolor": "1.1.1"
   },
   "devDependencies": {
     "@mp/tsconfig": "workspace:*"
