--- apps/game/server/area/schema.ts
@@ -1,4 +1,0 @@
-import { varchar } from "@mp/db";
-import type { AreaId } from "../../shared/area/area-id";
-
-export const areaId = () => varchar({ length: 60 }).$type<AreaId>();
