+++ apps/game/server/index.ts
@@ -2,17 +2,15 @@ export * from "./npc/npc-ai";
 export * from "./npc/npc-spawner";
 export * from "./npc/rpc";
 export * from "./npc/service";
-export * from "./npc/schema";
+export * from "./npc/types";
 export * from "./character/character-remove-behavior";
 export * from "./character/rpc";
 export * from "./character/service";
-export * from "./character/schema";
+export * from "./character/types";
 export * from "./game-state";
 export * from "./user/client-visibility";
-export * from "./service";
 export * from "./area/lookup";
 export * from "./area/rpc";
-export * from "./area/schema";
 export * from "./traits/combat";
 export * from "./traits/movement";
 export * from "./traits/appearance";
