+++ apps/loadtest/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "@mp/tsconfig/vite.json",
+  "extends": "@mp/tsconfig/node.json",
   "compilerOptions": {
     "outDir": "dist",
     "noEmit": false
