+++ apps/server/package.json
@@ -23,7 +23,6 @@
   },
   "dependencies": {
     "@mp/auth": "workspace:*",
-    "@mp/db": "workspace:*",
     "@mp/env": "workspace:*",
     "@mp/game": "workspace:*",
     "@mp/ioc": "workspace:*",
@@ -40,13 +39,17 @@
     "@mp/ws": "workspace:*",
     "cors": "2.8.5",
     "dotenv": "16.4.5",
+    "drizzle-orm": "0.38.3",
     "express": "4.19.2",
-    "import-in-the-middle": "1.12.0"
+    "import-in-the-middle": "1.12.0",
+    "pg": "8.13.1",
+    "unique-names-generator": "4.7.1"
   },
   "devDependencies": {
     "@mp/tsconfig": "workspace:*",
     "@types/cors": "2.8.17",
     "@types/express": "4.17.21",
+    "@types/pg": "8.11.10",
     "drizzle-kit": "0.30.1",
     "tsx": "catalog:"
   }
