+++ apps/server/src/db/client.ts
@@ -1,6 +1,8 @@
+import type { Pool } from "pg";
 import pg from "pg";
 import { drizzle } from "drizzle-orm/node-postgres";
-import type { DbClient } from "./client";
+
+export type DbClient = ReturnType<typeof drizzle<Record<string, never>, Pool>>;
 
 export function createDbClient(connectionString: string): DbClient {
   const pool = new pg.Pool({ connectionString });
