+++ apps/server/src/db/types/path.ts
