+++ apps/server/src/db/types/short-id.ts
