+++ apps/server/src/db/types/vector.ts
