--- apps/server/src/schema.ts
@@ -1,1 +1,0 @@
-export { npcTable, npcSpawnTable, characterTable } from "@mp/game/server";
