--- packages/db/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "@mp/db",
-  "version": "0.0.0",
-  "sideEffects": false,
-  "type": "module",
-  "license": "MIT",
-  "exports": {
-    ".": {
-      "types": "./src/index.ts",
-      "import": "./src/index.ts",
-      "require": "./src/index.ts"
-    },
-    "./server": {
-      "types": "./src/server.ts",
-      "import": "./src/server.ts",
-      "require": "./src/server.ts"
-    }
-  },
-  "dependencies": {
-    "@mp/math": "workspace:*",
-    "@mp/std": "workspace:*",
-    "drizzle-orm": "0.38.3",
-    "pg": "8.13.1"
-  },
-  "devDependencies": {
-    "@mp/tsconfig": "workspace:*",
-    "@types/pg": "8.11.10"
-  }
-}
