--- packages/db/readme.md
@@ -1,3 +1,0 @@
-# @mp/db
-
-Encapsulation of the orm parts of drizzle
