--- packages/db/src/client.ts
@@ -1,4 +1,0 @@
-import type { Pool } from "pg";
-import type { drizzle } from "drizzle-orm/node-postgres";
-
-export type DbClient = ReturnType<typeof drizzle<Record<string, never>, Pool>>;
