--- packages/db/src/index.ts
@@ -1,14 +1,0 @@
-export * from "drizzle-orm";
-export * from "drizzle-orm/pg-core";
-export * from "./client";
-export { vector } from "./types/vector";
-export { path } from "./types/path";
-export * from "./types/short-id";
-
-export type {
-  ColumnsWithTable,
-  SelectedFields,
-  SelectedFieldsFlat,
-  SelectedFieldsOrdered,
-  TableConfig,
-} from "drizzle-orm";
